backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
coverage==7.6.8
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.3
django-cleanup==9.0.0
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
drf-spectacular==0.27.2
drf-spectacular-sidecar==2024.11.1
hiredis==3.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.0
pytest==8.3.3
pytest-cov==6.0.0
pytest-django==4.9.0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.0.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.21.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
